Устранил замечания, поправил проблему с сохранением User'а в БД.
